Various related changes.  Only the first two will be noticed to anyone besides Gordon.
1) const-correctness
- Nrrd* function parameters which are input, in that they point to nrrds
which are only read and not written, are now const Nrrd *.
- For the time being, I have also changed Nrrd** inputs to
const Nrrd *const *, because this page says its okay:
<http://www.parashift.com/c++-faq-lite/const-correctness.html#faq-18.15>
However, I get warnings when I pass a Nrrd** as such an argument on darwin's
gcc.  One (annoying) way to avoid these warnings is to create a second array
of const Nrrd**, and copy all the Nrrd* pointers over.

2) NrrdRange, Nrrd
This is a new struct which contains only:
  double min, max;
  int hasNonExist
These fields have been REMOVED from the main Nrrd struct, because
this kind of meta-information can't be reliably kept correct, and because
those fields end up being used as input parameters to things like
nrrdQuantize and nrrdHistogram, in which case min and max have nothing
to do with the values in the nrrd, but with the desired action of the
nrrd function.  Such usage idioms stood in the way of const correctness
and being data-flow friendly.

3) When reading nrrds, the "min: " and "max: " fields are now
completely ignored in the same way that the old "number: " field is
ignored.  The "min: " and "max: " fields are now never written.

4) NrrdIter
now has a Nrrd*, and a const Nrrd*, depending on whether the NrrdIter
owns the nrrd or not.

4) There are various functions for creating NrrdRanges, most of which
allow explicit control over the min-max range for 8-bit data is
actually determined or simply assumed to be 0-255 for unsigned chars
(for example).  These are facilitated by the new nrrdBlind8BitRange*
enum values.

5) nrrdNonExist --> nrrdHasNonExist
The nrrdNonExist* enum values have been correspoding nrrdHasNonExist* names.

6) The nrrd-supplied NrrdKernels now have type NrrdKernel *const.
Perhaps it should be const NrrdKernel *const.

7) added nrrdTernaryOpAdd, nrrdTernaryOpMultiply, nrrdTernaryOpMin,
nrrdTernaryOpMax, and nrrdTernaryOpIfElse.


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@1702 3d70eeeb-363e-0410-a505-8a46323a89f2