All the input/output in nrrd has been reorganized.
It used to be a completely adhoc mess, now I'm almost proud of it.
There are no significant functional API changes: nrrdRead, nrrdLoad,
nrrdWrite, and nrrdSave still function as they used to, with the minor
exception described in the "BEHAVIOR CHANGE" note below.  I doubt
there are any nrrd users that will notice the minor API changes that
there are, so you can probably consider the rest of this as bedtime
reading.

The basic change is that there are two new structs in nrrd:
NrrdEncoding and NrrdFormat.  Instead of scattering all information
relevant information across various arrays of ints and functions, now
everything related to one encoding or format is centralized in one
place.  All the nrrd I/O functions use the NrrdEncoding and NrrdFormat
pointers they are given, whether or not the underlying structs are
supplied by nrrd or someone else.  Some may generously call this a
"plugin" architecture.  The NrrdKernel and gageKind structs have
been this way since the beginning.

BEHAVIOR CHANGE:
- nrrdRead/nrrdWrite no longer resets the given NrrdIO to initial state
  before returning, which means that you have to call nrrdIOInit() on it
  yourself if you want to reuse the NrrdIO between nrrd reads.
- (should have been done with advent of NrrdRange: the "min: " and
  "max: " fields of nrrd files are now ignored on reading (there's no
  more min and max fields in the nrrd), and never written
- nrrdSave/nrrdWrite can now save out EPS images.

API GONE:
- airEnum *nrrdMagic: formats are now themselves primarily responsible for
  recognizing files of their type.
- nrrdIOReset: use nrrdIOInit instead
- nrrdDefWriteSeperateHeader: stupid idea
- nrrdTable: no one cared
- int nrrdFormatIsAvailable[]
- int nrrdFormatIsImage[]
- int nrrdEncodingEndianMatters[]
- int nrrdEncodingIsCompression[]
- int nrrdEncodingIsAvailable[]
- nrrdFitsInFormat(): now use format->fitsInto
- nrrdDirBaseSet: hopefully obsolete now

API RENAME:
- In general what was called the "table" format is now called "text".
- int nrrdDefWriteBareTable --> nrrdDefWriteBareText
- airEnum *nrrdEncoding --> nrrdEncodingType:
- airEnum *nrrdFormat --> nrrdFormatType: these help in identifying any
  of the nrrd-supplied encodings or formats

API NEW:
- nrrdIOInit(): returns a NrrdIO to initial state
- NrrdFormat *nrrdFormatUnknown;
- NrrdFormat *nrrdFormatNRRD;
- NrrdFormat *nrrdFormatPNM;
- NrrdFormat *nrrdFormatPNG;
- NrrdFormat *nrrdFormatVTK;
- NrrdFormat *nrrdFormatText;
- NrrdFormat *nrrdFormatEPS;
- NrrdEncoding *nrrdEncodingUnknown;
- NrrdEncoding *nrrdEncodingRaw;
- NrrdEncoding *nrrdEncodingAscii;
- NrrdEncoding *nrrdEncodingHex;
- NrrdEncoding *nrrdEncodingGzip;
- NrrdEncoding *nrrdEncodingBzip2;
- nrrdFormatArray[], nrrdEncodingArray[]: arrays of all nrrd-provided
  formats and encodings.


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@1754 3d70eeeb-363e-0410-a505-8a46323a89f2