Nrrd (and the rest of teem) is supposed to obey a rule of thumb with respect to strings (char*) and and their allocation:
  Strings are always re-allocated, and can be free()ed without worry

char* pointers given to nrrdKeyValueAdd are airStrdup()d to make the
strings that are stored internally. But before now, nrrdKeyValueGet
and nrrdKeyValueIndex would return pointers to memory INSIDE the
nrrd, leading to problems if they were freed.  Bad!  Now, these
functions return NEWLY allocated copies of the strings, so that the
caller can (and should) free them.

However, power users may want to avoid this, so they can set
nrrdStateKeyValueReturnInternalPointers, which means that (as before)
internal pointers are returned.

Also, nrrdKeyValueCopy() was created, and now nrrdCopy() calls that
instead of copying key/value pairs by hand.


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@1951 3d70eeeb-363e-0410-a505-8a46323a89f2