API NEW, FILE FORMAT NEW:
This is a big change: Nrrd can now represent general orientation
information, beyond the simple axis-aligned position stuff that is
handled by the per-axis min/max/spacing.

I've chosen to put all the orientation information into a set of
per-axis vectors, and a single per-array vector.  All the new fields
in Nrrd have to do with identifying the number of coefficients in
those vectors, or describing the basis vectors and units with which
those coefficients are measured.  The per-axis vector is called
"spaceDirection", it is now a member of the NrrdAxisInfo struct.  ***
Despite how the name might imply a unit-length vector, the
spaceDirection vector in fact represents both the sample spacing along
that axis (recovered as the vector norm), and the direction of the
axis ***.  The spaceDirection vector for axis i identifies the
difference in spatial location between samples whose indices differ by
1 along axis i, extending from the low index to the high index
positions.  For example, spaceDirection for axis 0 is the vector from
the location of sample[0,0,0] to sample[1,0,0].  So the spaceDirection
vectors define the edges of the voxel.

The per-array vector is called "spaceOrigin", it now a member of the
Nrrd struct itself.  I (with Dave's input) decided to go with the name
"origin" because it seemed more precise than the term used in the
DICOM format, "location", and because VTK currently uses "ORIGIN" for
the same point.  spaceOrigin is a new member of the Nrrd struct.  Like
DICOM, this identifies the spatial location of the center of the
"first sample transmitted"- the sample with the lowest memory
location.

Here are the other new fields in the Nrrd struct.  The dimension of
the space that spaceDirection and spaceOrigin vectors live in (the
number of valid coefficients that they have) is called
"spaceDimension".  *** This is a separate variable from the intrinsic
array dimension ("dimension") because it logically is separate ***.
For example, a single time-step of an fMRI experiment is a 3D array
which lives on a slightly sheared slice in a 4D XYZT space, while a
DT-MRI scan is a 4D array which lives in a 3D XYZ space.  When known,
the canonical space containing the vectors is represented by "space",
which takes on integral values from the new "nrrdSpace" airEnum.
Finally, the units of measurement of the vector coefficients are
represented by an array of strings, "spaceUnits".  Here's how the
basic structs look now:

Associated with these new orientation variables are new fields in the
NRRD file format:

  space:
  space dimension:
  space units:
  space directions:
  space origin:

Using these fields requires using the new magic "NRRD0004".
nrrdLoad() and nrrdRead() can read these, and nrrdSave() and
nrrdWrite() and write them.  Here's how the new fields are used.

You can either use "space: " or "space dimension: " to describe the
space, but not both.  When you use "space: ", you're defining both
nrrd->space and nrrd->spaceDim.  Some samples of using this:

  space: RAS
  space: right-anterior-superior
  space: LAS
  space: left-anterior-superior
  space: LPS
  space: left-posterior-superior

"space: " are parsed with the nrrdSpace airEnum, so different strings
can be used to represent the same setting.  If someone wants to use
nrrd to represent orientation in a space which is as yet unnamed, they
can identify the spaceDim itself:

  space dimension: 3

Either "space: " or "space dimension: " must be used before any of the
other new fields, because these identify the number of coefficients in
the spaceDirection and spaceOrigin vectors that will be parsed.  All
these vectors are represented in a form like "(x,y,z)".  For example:

  space origin: (-10,-10,-20)

Then, like all per-axis fields, you give as many spaceDirection
vectors as there are axes:

  space directions: (1,0,0) (0,1,0) (0,0,2)

Finally, the units for each coefficients of these vectors is given as
a list of quoted strings:

  space units: "mm" "mm" "mm"

Note that if you wanted to represent a single fMRI timestep, where
each Z slice lives it a slightly different time, you could use:

  dimension: 3
  space: RAST
  space units: "mm" "mm" "mm" "ms"
  space origin: (-10,-10,-20,0)
  space directions: (1,0,0,0) (0,1,0,0) (0,0,2,0.01)

Note that you still have one spaceDirection vector for each of the
three array axes, but the number of coefficients in those vectors is
determined by the "space" (right-anterior-superior-time).

----------------- interaction with non-scalar types -----------------

When you use an array axis to represent non-scalar data, that axis
doesn't really have any spatial existence.  Just like the existing
per-axis min, max, and spacing, Nrrd uses NaN (not-a-number) to say
"this axis has no spaceDirection".  For example:

  dimension: 4
  kinds: nrrdKind3DMaskedSymMatrix nrrdKindDomain nrrdKindDomain nrrdKindDomain
  sizes: 7 256 256 90
  space dimension: 3
  space units: "mm" "mm" "mm"
  space origin: (-10,-10,-20)
  space directions: (nan,nan,nan) (1,0,0) (0,1,0) (0,0,2)

As a convenience, "none" can be used for an all-NaN vector:


  space directions: none (1,0,0) (0,1,0) (0,0,2)

The non-spatial vector is not included in the IJK-to-XYZ transform.
The AIR_EXISTS() macro is still used to determine if value is NaN
(just as its currently used for the per-axis min, max, and spacing).

----------------- interaction with axis-aligned position -----------------

Nrrd already has a simple notion of sample "position", based on the
per-axis min, max, and spacing.  This is used, for example, to
represent a VTK STRUCTURED_POINTS grid which is axis-aligned, but has
different spacings along each of the three axes:

  dimension: 3
  sizes: 256 256 80
  spacings: 1 1 3
  units: "mm" "mm" "mm"
  mins: -10 -10 -20

In this case, the per-axis "min" plays the role of the VTK "ORIGIN".
When an axis represents a histogram, min and max are used together.
For a joint histogram of two scalars, for example, you might have:

  dimension: 2
  sizes: 500 500
  mins: -10 0.3
  maxs: 40 0.9

The sample position information implied by min, max, and spacing is
inherently one-dimensional, with the assumption that the array axis is
somehow aligned with the axis of the enclosing space.

To enforce some minimal semantic consistency between axis-aligned and
general orientation information, there must be mutual exclusion
between the use of min/max/spacing/units, and the spaceDirection
vector.  That is, for each NrrdAxisInfo *axis in the array, you can
either use axis->min, axis->max, axis->spacing, and axis->units (or
some subset of those), or you can use axis->spaceDirection.  But using
axis->spaceDirection disallows using the other fields.

This exclusion does not apply to axis->thickness, axis->center (for
representing cell versus node centering), axis->kind, and axis->label,
because these aspects of per-axis meta-information do not constrain
anything about sample positions.  I initially thought that
node-versus-cell centering did effect sample positioning, but it turns
out it doesn't.

For a crazy example, here's an array storing per-voxel histograms of
RGB color values.  The histogram axis index is 1.

  dimension: 5
  space: RAS
  sizes: 3 100 256 256 80
  mins: nan 0 nan nan nan
  maxs: nan 255 nan nan nan
  space units: "mm" "mm" "mm"
  space origin: (-10,-10,-20)
  space directions: none none (1,0,0) (0,1,0) (0,0,1)

There is no problem here because axis 0 uses neither min/max or
spaceDirection, axis 1 uses min/max but not spaceDirection, and the
other axes use spaceDirections but not min/max.


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2368 3d70eeeb-363e-0410-a505-8a46323a89f2