[Mon Jun 19 05:31:14 CDT 2023] Much of the content in here had been essentially unchanged since revision 1327 2003-03-30 07:21:37 ("hello to the first windowed sincs"), and the tweaks that came in the following weeks (r1328, r1329, r1346), including r1347 that introduced using Taylor expansions zero 0, to avoid numerical instabilities.
A big bug originated in one of those early edits, wherein the _HANN macro (basis for
evaluating nrrdKernelHann) had an incorrect Taylor expansion, which meant that it
evaluated to 1.1 near 0, not 1.0; yikes!  Recent work of revisiting the various kernels
prior to a 1.13 release discovered this.

The math implementation of these kernels has been redone to fix that bug, and to make an
effort at improving numerical stability, now that GLK understand that topic better than
20 years ago. The Hann- and Blackman-windowed sinc kernels (nrrdKernelHann and
nrrdKernelBlackman) themselves did not actually need the near-zero Taylor approximation;
their numerical stability was fine. But the *derivatives* of these kernels do benefit
from a Taylor approximation near zero, which is used here. Further from zero, the
numerical stability of the derivative kernels is still pretty lousy compared
to the niceness of the other piece-wise polynomial kernels in nrrd.  For the usual
purposes of image resampling and reconstruction, that instability may not be the biggest
problem, but it should still be revisited and improved further.

The transition between the Taylor approximation and the regular evaluation (for the 1st
and 2nd derivatives of Hann and Blackmann) has been set up to be C0 continuous, as
follows.  The Taylor approximation is made of two terms, and is used whenever |x| is
below some threshold "CUT" (that does not depend on radius R, in contrast to
earlier attempts at this). The higher of the two Taylor terms is scaled (as a function of
R) so that the approximation and the correct formula agree exactly at CUT. A second
macro for that scaling function ("scl") is passed to the function bodies for all the
kernel methods. Getting that math right has benefited from an extra sanity check to be
added to the these kernels: the radius parameter (parm[1]) is now clamped from below to
be 2: you didn't really want a windowed sinc if you're not letting it ring at least once.

All the large math expressions are from manual massaging of copying from Mathematica.
In double-precision, the absolute relative error of the Taylor approximations (in the 1st
and 2nd derivatives of nrrdKernelHann and nrrdKernelBlackman) is less than 3e-8, but that
is the error bound for the minimum radius parm[1] of 2: the error goes down (the tweaked
Taylor approximation becomes more accurate) with a larger radius.

Some initial efforts were made to create versions that only compute in single-precision
float (with more type-specific arguments to macros, like cosf,sinf instead of cos,sin),
but even before that was finished, it was clear that the numerical quality was total
garbage, especially for the kernel derivatives - these formula are just so gangly that
there are too many occasions for rounding error for single-precision to work usefully.
Totally different approximation schemes (that do not ever rely on the analytic formulae)
may fare better.  Hence the kernel function bodies for the _1_f and _N_f methods don't
even bother pretending to work in single-precision, they just use double precision for
everything.


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@6953 3d70eeeb-363e-0410-a505-8a46323a89f2