Having revisited hest code with prior change and bug fix, was now able to see that support --help was pretty easy; it can be done inside copyArgv (with previously inscrutable name _hestResponseFiles). So --help support is finally done:
The FIRST hestOpt add in an hestOpt array will have its new helpWanted field
set, if "--help" was a command-line argument, and if
hestParm->respectDashDashHelp is true (which it is by default).  In that case,
hestParse() does not return non-zero: it indicates a successful parse, even
though nothing has actually been parsed.  With --help, there is no
command-line parsing error, and the error string has not been set.

NOTE: This is effectively an API CHANGE for how people call hestParse: they
have to not just check for non-zero return but also (with zero return)
opt[0].helpWanted.

Also:
- Removed all but one "hammerhead" reference, and explained it more
  in a comment.
- Completely re-write hestParseOrDie which had become too messy to
  clealy add new handling of --help; makes more sense now. But the intent
  is that functionality is the same.


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@6994 3d70eeeb-363e-0410-a505-8a46323a89f2