Two changes in preparation for non-var-arg alternatives to hestOptAdd:
(1) (a change in the API but not a breaking API change)
The hestOpt struct gains some long-overdue const-correctness; these fields
   airEnum *enm;       /* used ONLY for airTypeEnum options */
   hestCB *CB;         /* used ONLY for airTypeOther options */
are now:
   const airEnum *enm;
   const hestCB *CB;
This isn't actually related to non-var-arg alternatives to hestOptAdd, but
was noticed as a drawback.

(2) API NEW: A new non-var-args version of hestOptAdd:
unsigned int hestOptAdd_nva(hestOpt **optP, const char *flag,
                            const char *name, int type, int min, int max,
                            void *valueP, const char *dflt, const char *info,
                            unsigned int *sawP, const airEnum *enm,
                            const hestCB *CB);
In fact, now hestOptAdd is implemented as a wrapper around this.
and the internals of methodsHest.c have been re-organized to facilitate this.



git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@7026 3d70eeeb-363e-0410-a505-8a46323a89f2