Now that we've already done an API CHANGE with hestOpt->helpWanted, another API CHANGE added two more things to hestOpt:   unsigned int arrAlloc and arrLen. (also reformatted many comments in hest.h)
These allow the hestOpt array to be managed like an airArray: it knows its own
allocated size (arrAlloc) as well as the logical length (arrLen), so that
re-allocation happens when the requested length exceeds some multiplement of
an allocation increment.  hestOptAdd() used to do clever things to manage the
hestOpt array with a terminator hestOpt, but was totally inscrutable; now it
is nicely explicit.  The new private methods optarr_init() and optarr_incr()
do the work.

API NEW: hestOptSingleSet sets a single hestOpt struct, with no var-args
And now hestOptAdd_nva relies on hestOptSingleSet for setting hestOpt members.

Thanks to this rethink, hestOpt->kind is no longer set by hestParse,
but as a side-effect of hestOptAdd.

Also, added some assert()s which simplify error handling of heap allocations
that are unlikely to fail.


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@7028 3d70eeeb-363e-0410-a505-8a46323a89f2