[Backport PATCH] COMP: -Werror causes build failure on security warning
Corresponding patch (0001-COMP-Werror-causes-build-failure-on-security-warning.patch)
was sent by email to Gordon L. Kindlmann <glk@uchicago.edu> on 2025-06-19 for
considered for integration into the upstream project.

---

A compiler warning related to a common security vulnerability
was issued as an error due to the use of the hard-coded -Werror flag.

The failure occurs on newer compilers available on Ubuntu 24.04.  The
issue can be fairly easily fixed by using snprintf, and some truncation
of the concatenated messages.