The new implementation of hestParse, currently called hestParse2, seems to finally be working.  The old hestParse remains, although with some of the changes below it may leak memory in ways it previously didn't.  After further debugging it will be removed (so parseHest.c) will be much shorter.
- ABI CHANGE: since hest's birth, the hestOpt struct has had an `int alloc`
that helped keep track of what was allocated as a result of parsing, and what
had to be freed.  It could take on magic values 1, 2, 3, or 4, and the logic
that hestParseFree had to implement to keep track of all this (in a way that
depended on the option type, and the non-NULL-ity of CB->destroy in the case
of airTypeOther) was just hideous.  Instead, now a `airArray *parseMop` is
used, to remember what needs to be freed.  Free'ing or destroy'ing callbacks
added here (by _hestParseSingle[type]) with when=airMopAlways. And now,
hestParse(2) just adds hestParseFree to its in-case-of-error mop. This is so
so much	simpler and easier than the original code.

- API CHANGE: hestParseFree now returns the the given `hestOpt *hopt`, instead
of void. This facilitates its use in mops.

- API CHANGE: removed the void*-taking-and-returning hestParmFree_vp and
hestOptFree_vp. These were experiments to make their use with an airMop less
clunky, because you no longer need the `(airMopper)` cast, but this was a
dumb idea: adding these _vp variants is work, so hundreds of other functions
in Teem do not have such a _vp variant, plus the `(airMopper)` cast idiom
is well-established.

- in many functions renamed parameter `hestOpt *opt` to `hestOpt *hopt`
for consistency with other hest parameters and variables


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@7520 3d70eeeb-363e-0410-a505-8a46323a89f2